Extract XML, XML Editor, and Cron Job tests
